refactor(-): :hammer: move viewset solicitacao medicao inicial
